BlogPost/requirements.txt
Django==3.0.3
djangorestframework==3.12.2
django-filter==2.4.0
django-rest-auth==0.9.5
requests==2.25.0
django-rest-multiple-models==2.1.3
djangorestframework-filters==0.11.1